Remove bogus parts of IUIElement interface
As an external, abstract interface, the constructor doesn't do
anything.  The members in it were parts of the base class
shaka.ui.Element, which implementations of IUIElement can still
inherit from.

In fact, there is not much reason to have this empty IUIElement
interface now, but we'll keep it for now to avoid breaking compiled
applications.

Change-Id: I5687b281be523c29656b5c51e63628ba164cf1b3