Prefer original source stream when matching for period flattening.
When we construct final streams in period flattening, we start them
off original streams and then combine streams from each period that
best matches the characteristics of the original stream.
The original stream also participates in the matching process, but
it was possible for it not to get selected as the best match if
there were other streams with similar characteristics.
This change adds logic to always prefer the original stream if
it's encountered.

Also, this change adds a bandwidth comparison for audio
streams to ensure the closest streams get matched bandwidth-wise.

Issue #2537.

Change-Id: I5d99a452a12ea1c4bb70fa8aba4bec479ff9d27b