Fix module wrapper to prevent conflicting exports
The module wrapper expects to be in charge of exporting to module
loaders, but so does eme-encryption-scheme-polyfill, which is compiled
into the bundle.  To resolve this, the wrapper can hide the "module"
symbol from the inner part of the bundle, preventing it from doing its
own exports.

In v2.5.x, this was hidden by an apparent compiler bug that resulted
in module.exports being renamed in the compiled bundle.  Since "Unfork
Closure base and update the compiler", Change-Id
I62ec61e82d4edf342b2c576c2d4f89f11562ee65, this is no longer the case.

Closes #2549

Change-Id: Id26002bc1c51ed423d77e7b3e7daf40b714b9f2f