Generate test content with multiple tracks
The test content generator was only designed for single-track content.
In the one remaining case where we generated multi-track content,
there was one-off logic to generate that content instead of generating
it based on descriptive metadata.

Now we can generate test content with multiple tracks from descriptive
metadata.  The tracks will still all contain the same thing, but the
metadata can be varied to create distinct resolutions and languages.

This is an important step toward generating content to build a
regression test for #2553.

Change-Id: Ic752ff1e50273470d2784910251d181fe995128a