Rename fetchPartialSegment to fetchStartofSegment
In HlsParser, fetchPartialSegment_() is meant to fetch the starting part
of the segment to get the segment's start time, so we don't have to
download the full segment if possible.
PART_SEGMENT_SIZE is the default size of the starting part of the
segment we fetch.

Low Latency HLS introduces 'Partial Segment' as a smaller part of
a regular segment, such as CMAF chunks, and a regular segment might be
divided into many partial segments.

Renaming the original function and field to improve code readability and
avoid any future confusion.

Issue #1525

Change-Id: I593497f6e99e9ffe6128d6976eb7886bbd79679c