Make the controls background gradient proportional
Rather than making the gradient 92px tall, make it 15% of the size of
the video container.  This is consistent with the size of the controls
and the position of the subtitles when the controls are shown.

In this way, the size of the gradient is now proportional and
responsive to the size of the video container.

This was spotted while developing screenshot-based layout tests using
relatively small containers.

Change-Id: I304d99a26be90b9eefa50dc179b924da15dc601d