fix(ui): Set an explicit font size for icons
We depended on this font size being set in the font server's CSS, but
did not realize we had this implicit dependency.  This makes our
desired font size explicit.

This protects us from applications which happen to use the same font
and class name with a different default size.  Note that applications
which intend to change our icon size can still do so explicitly in
their own styles.  This is meant only to protect against accidental
conflicts.

Closes #2633

Change-Id: I0a2993c8b2e39596b47eb59b7e1a4a6373633c0b