fix: Correct license headers in CSS/LESS
The LESS tool respects the same "/*!" syntax for forcing the inclusion
of a license header, but will not dedup these licenses.  So the LESS
files generally will not use this syntax.  Instead, the build system
prepends a license header after compilation.

The exception is for our SVG spinner, which is based on third-party
CSS from codepen.  The copyright header for this is forced into the
output to give proper credit to the original author.

Issue #2638

This is a minimal backport to v2.5.x which only touches the header in
the output.

Change-Id: I4c58e2b082f2d5e550a6f0a30feaaf9ebf82a53a