fix(demo): Fix missing icon for demo app menu
This icon went missing in the change "Apply upstream styles for icons
font" (Change-Id: I615c1dfd86e5b9802d03050c5db0154ac196c2ef).  We may
not be using that CSS class name any more, but the MDL library still
is.  So this override that was removed needed to come back, albeit for
a different reason than before.

Change-Id: Ibf7e865604f09d9d02fe15fe9f577ea63cdf6d1d