test: Detect screenshot support via WebDriver
We have added new platforms to our Selenium grid which do not support
screenshots.  There is no direct way to detect this support without
taking a screenshot, so this change will take a screenshot the first
time a test wants to know if they are supported.  The result will be
cached per browser to improve performance.

Change-Id: I30217db91b10336efee454c4781782052bc17572