fix: Re-add a session close workaround.
There is a new bug that is preventing the DRM engine from
shutting down, for Widevine assets. Or, at least, new to us.
It happens if session.generateRequest is called, but
session.update is never called (e.g. if the network request fails).
In this case, session.close will never resolve.
This re-adds the old workaround we used to use, where we
would timeout after a while, if the session fails to close.

This also adds a regression test.

Issue #2741

Backported to v2.5.x

Change-Id: Ieb6a64b762473edb9a3209580a84aee870e21fb5