fix(Tizen): Fix initial bandwidth estimate on Tizen
Tizen 3 has the NetworkInformation API, but not the downlink
attribute.  This led to NaN as a bandwidth estimate.  Now we screen
for the presence of this attribute as well as the API overall.

Futhermore, we now disable this API in unit tests to better control
the fake estimates used during the tests.

Found while investigating #2620.  Introduced in PR #2663, and has not
affected any release versions.

Change-Id: Iaa1486545825ceee536ecbe5ea617f92de4fbc2d