fix(Tizen): Fix src= playback on Tizen 3
src= playback on Tizen 3 was stalled at time 0, causing automated test
failures.  These went unnoticed while our Tizen device in the lab was
offline.

The only workaround I was able to find was to call load(), which we
recently stopped doing for src= playback.  This adds it back, but only
for Tizen.

Found while investigating #2620.

Backported to v2.5.x

Change-Id: I4e071a6f1ccb18b9891dfc646a22dbb8c1711eb4