test: Update Karma and Jasmine
The latest version of Jasmine has deprecated several properties of
jasmine.Env in favor of a configure() method.  This updates our
bootstrap code to use the configure() method.

The latest version of the karma-jasmine adapter hard-codes its own
specFilter configuration at a time that overwrites ours.  There is no
way to configure karma-jasmine not to do this, so we shim
jasmine.getEnv().execute() to work around it.

I filed karma-runner/karma-jasmine#273 to track this latter issue.

Change-Id: Ie6a32f67009b52edeeb8f1974ee131753a75ba3f