Support prefetch of mpeg-dash segments
This diff adds the ability to download segments listed in a
mpeg-dash manifest in parallel. Before this diff, the segments
would be downloaded serially. This despite knowing the list of
segments to be downloaded apriori. This diff changes that allows
shaka-player to prefetch segment downloads.

The exact level of parallelism, or the exact number of segments
to be prefetched is controlled via a configuration "prefetchLimit".
When set to zero, it defaults back to older behaviour. When set
to a positive number shaka-player would prefetch only those many
segments in parallel and then wait until some of the ongoing
downloads finish before issuing more prefetch requests.

There are some strange observations w.r.t. behaviour with range
requests, pre-flight OPTIONS request and chrome. In many cases
every segment request is preceeding by a OPTIONS request. With
browser's cache disabled this often leads to multiple identical
OPTIONS request. When cache is enabled, that causes an undesirable
behaviour with chrome. Chrome's cache has a writer lock and it blocks
any subsequent request that has the same URL as the one being
currently downloaded. This for range requests often defeats the
purpose of the diff as it leads to serialization of all range
requests for the same segment (same URL). Thankfully, firefox
does not see this issue.

Tests:
o Ran this version of shaka-player on my local setup and saw
  segments getting prefetched in parallel.
o By changing the prefetchLimit, checked that the number of
  segments downloads being in flight mimicked the configuration.
o Tested the same on chrome as well as firefox.