fix(period flattening): Ensure the number of variants stays stable when new periods are added.
Our period flattening logic used to keep creating new streams and variants
when encountering a certain type of content (same main content and slightly differing
sequence of ads). In time, this led to a performance drain, where more and more time
was spent processing the variants, leading to all kinds of delays.
This is the last change in the series, addressing the way we set stream properties
when concatinating streams.

Fixes #2716.
Fixes #2736.

Change-Id: I98ae692797d19344601d7a0c51d1e343dfe55e01