fix(text): Fix ordering of cues on IE and Edge
An old fix for an ordering issue on modern browsers was creating the
opposite effect on IE11 and legacy Edge, and caused cues to be shown
in the wrong order.  This is because the original issue only affects
native VTTCue, which IE11 and legacy Edge don't have.

This fixes the issue by checking to see if we are using native VTTCue
before flipping the order of cues with identical times.

This issue was caught by new native text display layout tests.

Change-Id: Id22d9c0da1580ab5b04664c6f584f10c0a787736