test: Fix test failures triggered with --no-babel
When we run ./test.py --no-babel, Babel is not used to transpile the
tests.  (This restricts test runs to modern browsers and prevents
testing on, say, IE11.)

But Babel seems to have masked some issues with constructor spies.
Jasmine will call your fake callback with "new" if the spy itself is
called with "new", but this means that constructor spies must have
plain function callbacks, not arrow functions.

Backported to v2.5.x

Change-Id: I936b024504c4173cc29824408e97e5c7d6c591d8