test: Fix native text layout tests on IE/Edge
The native layout tests on IE & Edge were hanging.  This seems to have
been caused in part by peculiarities of the video stream we used in
the tests.

This replaces the video, which was 10s of 32x32 solid green, with
another which is 320x180.  The new file is only 9kB, so it's not much
larger than the old one.

The old video would somehow hang IE & Edge, seemingly due to its
unusual tiny frame size.  I just tried a larger frame somewhat
arbitrarily and found that it worked.  I have no model for why this
would be the case.

This also makes the tests somewhat more robust by seeking to the
beginning before a "wait for movement" step in the test setup.  This
seek to 0 fixed part of the problem early on, before I replaced the
video.  I don't know if it's strictly necessary in addition to
replacing the video, but I did it first and it makes logical sense to
do it, so I've left it in the commit.

Finally, the new video has the same aspect ratio as the screenshots
themselves, and completely fills the frame.  This was not true before,
which meant that the background color on the left and right sides
didn't exactly match the video frame color in the middle.  (And on IE,
the video frame color wasn't showing in the screenshots at all for
some reason.)  Now that the video has been replaced with something
that seems to behave better on IE & Edge, the screenshots also need to
be updated to match on all platforms.

Fixed: 165712112

Change-Id: I6156366b230b71fc52e095b62dd51be8f3e68b93