test(LL-HLS): Parse preload hint segments
Add tests for segments with EXT-X-PRELOAD-HINT tags in Manifest Parser.
Preload hint init segment:
EXT-X-PRELOAD-HINT:TYPE=MAP,...
The ManifestParser should set the tag's name to EXT-X-MAP, so that we
can handle it the same way as regular MAP tags in the HlsParser.

Preload partial segment:
EXT-X-PRELOAD-HINT:TYPE=PART,...
The ManifestParser should put it into the partial segments list, so that
we can handle it similarly to a parital segment.
A partial segment has duration info and a preload hinted segment
doesn't, because the preload hinted segment is not finished yet.

Issue #1525

Change-Id: Ie733ee367cc639f8ff18ffb4f53bc7f1207c291e