fix(demo): Suppress bad errors when downloading.
When you pressed the download button, the console would
be filled with pointless assertion failures, when in
uncompiled mode.
This was caused by a bad assertion, that assumed that the
action bar on an asset card contained only HTMLButtonElement
objects.
This was true at the time the download button was first added,
but since then the tooltip system has changed, so there are
now some tooltip-related elements within that div also.
Thus, the assert has changed into a conditional, instead.

Change-Id: I34ce197dbd76c5e7c0e233001c18b197cce1803c