fix(drm): Fix passing sub-views through EME.
This fixes our handling of sub-views when passing init data through to
EME.  This is used by our offline code to pass the PSSH data from the
init segments.  By passing the ".buffer" field, we were passing the
whole underlying buffer and not the portion we cared about.

Issue google/shaka-player-embedded#158

Change-Id: Ia295d823391dc59f39f1019a03fe2f6da32139ff