fix(text): Fix bug where cue comparison throws
The cue comparison method references "this.equal",
which worked on the initial invocation of the method
but would fail on recursive calls, as "this" was
dereferenced along the way.
This was causing the cue comparison method to fail
between cues that had nested cues.
This changes the reference to be a static reference.

Change-Id: I1b8829987e3608db449213d7aafc84604fc7edac