chore(compiler): Prep for future compiler upgrade
An upcoming compiler release will use console.error in the Promise
polyfill.  This happens if the polyfill is unable to dispatch an
"unhandledrejection" event when a Promise without a rejection handler
is rejected.

This change explicitly allows the use of console.error in the
polyfill, so that the compiler upgrade will not break our build.

Change-Id: Ia0b5bc52a1a638937f4c952512f1a826476a3ab9