fix: Fix rapid seeking leading to infinite buffering
When we clear the buffer on a seek, there is a race between buffering
new data and clearing the buffer.  If data is already being fetched,
we need to abort the request when we clear the buffer.  Otherwise, we
could end up buffering at the wrong position in the stream, which in
some cases can lead to a buffering state that only ends when we seek
again.

Closes #2670

Change-Id: I8cde52734c282d188c5b3af6358d6b47c30ec3cd