fix: Fix non-deterministic exception in StreamingEngine
Due to the async nature of creating a segment index, it is possible to
enter an update cycle with a segment index in the stream, but no
segment iterator in the media state.  This was causing an exception
later on in getSegmentReferenceNeeded_ ("Cannot read property 'seek'
of null").

The fix is to clean up and separate the creation of segment index and
segment iterator.  We can't assume that already having a segment index
means we will also have a matching segment iterator.  Instead of
creating the segment index in switchInternal_, it will now be created
only before use (by update_ and by makeAbortDecision_), and always
without regard for the prior state of the segment index.

Fixes #2768

Change-Id: I043d860efe4b6f5fd6bbe16b4ff3e674ba46c7ae